Move footpath function to ui
