+++ src/openrct2-ui/libopenrct2ui.vcxproj
@@ -73,6 +73,7 @@
     <ClInclude Include="interface\Theme.h" />
     <ClInclude Include="interface\Viewport.h" />
     <ClInclude Include="interface\ViewportInteraction.h" />
+    <ClInclude Include="interface\ViewportQuery.h" />
     <ClInclude Include="interface\Widget.h" />
     <ClInclude Include="interface\Window.h" />
     <ClInclude Include="ride\Construction.h" />
@@ -135,6 +136,7 @@
     <ClCompile Include="interface\Objective.cpp" />
     <ClCompile Include="interface\Theme.cpp" />
     <ClCompile Include="interface\ViewportInteraction.cpp" />
+    <ClCompile Include="interface\ViewportQuery.cpp" />
     <ClCompile Include="interface\Widget.cpp" />
     <ClCompile Include="interface\Window.cpp" />
     <ClCompile Include="ride\Construction.cpp" />
