+++ src/openrct2/world/Footpath.h
@@ -191,8 +191,6 @@ money64 FootpathProvisionalSet(
     PathConstructFlags constructFlags);
 void FootpathProvisionalRemove();
 void FootpathProvisionalUpdate();
-CoordsXY FootpathGetCoordinatesFromPos(const ScreenCoordsXY& screenCoords, int32_t* direction, TileElement** tileElement);
-CoordsXY FootpathBridgeGetInfoFromPos(const ScreenCoordsXY& screenCoords, int32_t* direction, TileElement** tileElement);
 void FootpathRemoveLitter(const CoordsXYZ& footpathPos);
 void FootpathConnectEdges(const CoordsXY& footpathPos, TileElement* tileElement, int32_t flags);
 void FootpathUpdateQueueChains();
